etl/requirements.txt
alembic==1.7.6
aniso8601==9.0.1
attrs==21.4.0
click==8.0.4
Flask==2.0.3
Flask-Migrate==2.5.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.1
greenlet==1.1.2
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
numpy==1.22.2
pandas==1.4.1
protobuf==3.19.4
psycopg2==2.8.4
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
